Login and Register Modal
No form validation yet. But the global state management for these components are done. has a 'sign in' on register modal and a 'create account' on login modal so user can either sign in if they have an account or create an account if they do not have one yet.